Combine middleware, fix name mismatch from ES
This commit folds the responseSetsFromSearch middleware into the
extractFromSearchResults middleware.

Also fixes an issue where question content was not showing up on the
survey show page because 'content' is called 'name' in ES.
